Refactor: Reorganized package structure.
