+++ src/main/java/akkynaa/moreoffhandslots/client/input/KeyBindings.java
@@ -1,5 +1,6 @@
-package akkynaa.moreoffhandslots;
+package akkynaa.moreoffhandslots.client.input;
 
+import akkynaa.moreoffhandslots.MoreOffhandSlots;
 import com.mojang.blaze3d.platform.InputConstants;
 import net.minecraft.client.KeyMapping;
 import net.neoforged.neoforge.client.settings.KeyConflictContext;
