+++ src/main/java/akkynaa/moreoffhandslots/network/handler/OffhandCycleHandler.java
@@ -1,6 +1,7 @@
-package akkynaa.moreoffhandslots;
+package akkynaa.moreoffhandslots.network.handler;
 
 
+import akkynaa.moreoffhandslots.network.payload.CycleOffhandPayload;
 import net.minecraft.world.InteractionHand;
 import net.minecraft.world.entity.player.Player;
 import net.minecraft.world.item.ItemStack;
